feat(readme): Revamp README for enhanced clarity and detail
- Expanded project overview to highlight key features and functionalities.
- Added comprehensive usage guide with command examples for namespace and Pokémon management.
- Included technical details on database schema and state management.
- Improved formatting and organization for better readability.

These updates provide users with a clearer understanding of the OxideBox project and its capabilities, facilitating easier onboarding and usage.

+++ oxidebox/src/cli.rs
@@ -19,11 +19,11 @@ pub enum Commands {
     Summon {
         namespace: String,
         name: String,
-        level: u32,
-        hp: i32,
-        attack: u32,
-        defense: u32,
-        speed: u32,
+        level: u8,
+        hp: u16,
+        attack: u16,
+        defense: u16,
+        speed: u16,
         pokemon_type: PokemonType,
     },
     Start {
